Move `ConnChecker` to `helpers.rs`
